# docs/api/openapi.yaml
swagger: '2.0'
info:
  title: netobs unified API
  version: 2.5.6
paths:
...
  /api/v1/node/{node}/pods:
    get:
      description: '노드에 스케줄된 pod 별로 신원 (namespace 와 pod 와 uid) 과 CPU (limit 대비 percent
        와 절대량 cores), memory (limit 대비 percent 와 working set bytes), network bytes
        rate (netobs pod bytes 의 5분 rate 합산) 와 상태를 한 응답으로 돌려준다. percent 는 limit 분모라
        limit 없는 pod 는 생략되고 절대량 필드로 표현된다 (#328 과 동일 규약). severity 는 pod pressure score
        3종 최대의 환산 (low/elevated/high) 이고, 종료 pod 와 미관측 사유 (unobserved_reason, no_traffic
        포함 #342) 는 pods API 의 규약 (#314, #320) 을 그대로 쓴다. 관측 판정 소스는 rate 결과 존재라 (pods
        API 는 시리즈 존재) 카운터 샘플이 부족한 신규 pod 의 짧은 창에서는 여기서만 일시적으로 사유가 붙을 수 있고 1분 내 자연
        수렴한다. 정렬은 namespace 와 pod 사전순으로 결정적이며 이상 우선 정렬은 severity 를 근거로 표시 계층이 수행한다.
        at 파라미터로 사건 시점을 재구성할 수 있다.'
      produces:
      - application/json
      tags:
      - interference
      summary: 노드 하위 pod별 자원 사용량 목록
      parameters:
      - type: string
        description: 노드 이름 (DNS-1123 형식)
        name: node
        in: path
        required: true
      - type: string
        description: 평가 시점 (RFC3339 또는 unix seconds, 생략 시 현재)
        name: at
        in: query
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/internal_correlation_api.NodePodsResponse'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/netobs_internal_apicommon.ErrorBody'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/netobs_internal_apicommon.ErrorBody'
...
  /api/v1/pod/{namespace}/{pod}:
    get:
      description: 단일 pod 의 기본 정보 (uid 와 node 와 pod_ip 와 owner) 와 차원별 health 3종 (pod
        단위 pressure score 의 1 - score 환산), vitals (limit 대비 CPU 와 memory 사용률과 working
        set, CPU 절대 사용량), cgroup CPU throttle 상세 (throttled 비율과 5분 throttled/total
        periods 와 limit cores, cadvisor 가 노출하는 클러스터에서는 quota 와 period), 네트워크 종합 (pod
        입도 재전송과 drop rate, 최대 smoothed RTT) 을 한 응답으로 합성한다. cpu_percent 는 limit 대비
        비율이고 cpu_usage_cores 는 5분 rate 의 절대량 (cores) 이라 limit 없는 pod 는 cores 만 존재한다
        (#328, memory 의 working set 절대값과 대칭). RTT 는 분위수가 아닌 최대값이라 필드명이 max_srtt_seconds
        다. 미관측 pod 는 필드 생략과 summary 사유로 graceful 처리한다.
      produces:
      - application/json
      tags:
      - interference
      summary: pod 단위 종합 상세
      parameters:
      - type: string
        description: pod 네임스페이스 (DNS-1123 형식)
        name: namespace
        in: path
        required: true
      - type: string
        description: pod 이름 (DNS-1123 형식)
        name: pod
        in: path
        required: true
      - type: string
        description: 평가 시점 (RFC3339 또는 unix seconds, 생략 시 현재)
        name: at
        in: query
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/internal_correlation_api.PodDetailResponse'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/netobs_internal_apicommon.ErrorBody'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/netobs_internal_apicommon.ErrorBody'
  /api/v1/pods:
    get:
      description: '파드별 namespace, 이름, uid, pod IP, host IP, node, workload(created_by),
        priority, phase, qos, 관측 커버리지(observed)를 kube-state-metrics 기반으로 돌려준다. observed
        는 netobs 의 eBPF 시리즈가 존재하는지로, false 면 unobserved_reason 이 미관측 사유를 분류한다 (#320,
        #342): agent_absent 는 노드에 netobs 미배치, host_network 는 pod IP 가 host IP 와 같아
        IP 귀속이 성립하지 않는 상태로 cgroup 힌트가 학습되면 live 로 전환될 수 있는 시점 의존 분류다. no_traffic 은
        agent 의 소켓 스캔이 pod netns 의 무소켓을 증명한 상태 (unix socket 전용 pod 등) 로 시리즈 부재가 정상이고,
        no_data 는 관측 가능한데 시리즈 부재다. 종료 pod (phase Succeeded/Failed) 는 telemetry 부재가
        정상이라 observed=false 에 사유가 생략된다 (#314). ?namespace 로 필터한다. 다른 API의 src_namespace/src_pod/pod_uid와
        동일 키로 매핑한다.'
      produces:
      - application/json
      tags:
      - inventory
      summary: 파드 인벤토리
      parameters:
      - type: string
        description: namespace 필터 (생략 시 전체)
        name: namespace
        in: query
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/internal_correlation_api.PodsResponse'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/netobs_internal_apicommon.ErrorBody'
...
definitions:
...
  internal_correlation_api.NodePodUsage:
    type: object
    properties:
      cpu_percent:
        type: number
      cpu_usage_cores:
        type: number
      memory_percent:
        type: number
      memory_working_set_bytes:
        type: number
      namespace:
        type: string
      network_bytes_per_sec:
        type: number
      phase:
        type: string
      pod:
        type: string
      severity:
        type: string
      uid:
        type: string
      unobserved_reason:
        type: string
  internal_correlation_api.NodePodsResponse:
    type: object
    properties:
      generated_at:
        type: string
      node:
        type: string
      pods:
        type: array
        items:
          $ref: '#/definitions/internal_correlation_api.NodePodUsage'
      summary:
        type: string
...
  internal_correlation_api.PodCpuDetail:
    type: object
    properties:
      limit_cores:
        type: number
      period_microseconds:
        type: number
      quota_microseconds:
        type: number
      throttled_periods_5m:
        type: number
      throttled_ratio:
        description: ThrottledRatio 는 최근 5분 CFS period 중 throttle 된 비율 (0-1) 이다.
        type: number
      total_periods_5m:
        type: number
  internal_correlation_api.PodDetailResponse:
    type: object
    properties:
      cpu:
        $ref: '#/definitions/internal_correlation_api.PodCpuDetail'
      created_by_kind:
        type: string
      created_by_name:
        type: string
      generated_at:
        type: string
      health:
        description: 'Health 는 차원별 health (0-1, 1 이 healthy) 다. pod 단위 pressure score
          rule 의 1 - score

          환산이며, score 미산출 차원은 엔트리가 생략된다.'
        type: object
        additionalProperties:
          type: number
          format: float64
      namespace:
        type: string
      network:
        $ref: '#/definitions/internal_correlation_api.PodNetworkDetail'
      node:
        type: string
      pod:
        type: string
      pod_ip:
        type: string
      summary:
        type: string
      uid:
        description: 기본 정보는 kube_pod_info 라벨에서 온다. pod 미관측 시 생략된다.
        type: string
      vitals:
        $ref: '#/definitions/internal_correlation_api.PodVitals'
  internal_correlation_api.PodInventory:
    type: object
    properties:
      host_ip:
        type: string
      namespace:
        type: string
      node:
        type: string
      observed:
        description: 'Observed 는 eBPF 관측 커버리지다 (#248). netobs 가 상시 수집하는 netobs_pod_bytes_total

          시리즈가 이 pod 에 존재하면 true, 없으면 no-data 로 판정한다.'
        type: boolean
      phase:
        type: string
      pod:
        type: string
      pod_ip:
        type: string
      priority_class:
        type: string
      qos_class:
        type: string
      uid:
        type: string
      unobserved_reason:
        description: 'UnobservedReason 은 미관측 사유 분류다 (#320, #342). agent_absent (노드에
          netobs 미배치),

          host_network (pod IP 가 host IP 와 같아 IP 귀속이 성립하지 않음), no_traffic (netns 소켓

          테이블이 비어 네트워크 미사용이 증명됨, unix socket 전용 pod 등), no_data (관측 가능한데

          시리즈 부재) 로 구분한다. host_network 는 확정이 아닌 시점 의존 분류로, cgroup 힌트 (#228)

          가 TCP 트래픽에서 학습되면 observed=true (live) 로 전환될 수 있다. observed=true 와

          종료/Unknown phase 는 생략된다.'
        type: string
      workload_kind:
        type: string
      workload_name:
        type: string
...
  internal_correlation_api.PodNetworkDetail:
    type: object
    properties:
      drop_per_sec:
        type: number
      max_srtt_seconds:
        type: number
      retrans_per_sec:
        type: number
  internal_correlation_api.PodVitals:
    type: object
    properties:
      cpu_percent:
        type: number
      cpu_usage_cores:
        type: number
      memory_percent:
        type: number
      memory_working_set_bytes:
        type: number
  internal_correlation_api.PodsResponse:
    type: object
    properties:
      generated_at:
        type: string
      pods:
        type: array
        items:
          $ref: '#/definitions/internal_correlation_api.PodInventory'
...
  netobs_internal_apicommon.ErrorBody:
    type: object
    properties:
      error:
        $ref: '#/definitions/netobs_internal_apicommon.ErrorDetail'
  netobs_internal_apicommon.ErrorDetail:
    type: object
    properties:
      code:
        type: string
      message:
        type: string
...